Backend/requirements.txt
asgiref==3.7.2
black==24.1.1
certifi==2023.7.22
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.10.0
defusedxml==0.7.1
Django==5.0.1
djangorestframework==3.14.0
fonttools==4.47.2
idna==2.10
kiwisolver==1.4.5
matplotlib==3.8.2
mypy-extensions==1.0.0
numpy==1.26.3
opencv-python==4.9.0.80
opencv-python-headless==4.8.0.74
packaging==23.2
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.2.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-magic==0.4.27
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-toolbelt==1.0.0
roboflow==1.1.18
scipy==1.12.0
six==1.16.0
sqlparse==0.4.4
supervision==0.18.0
tomli==2.0.1
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.2.0
